+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release 1.16.0
 
+## General
+- Improved startup time
+  - Optimized registering RPCs (up to ~2700% faster) and online states (up to ~100% faster)
+  - Total Rain Meadow load time up to ~20% faster
+
+### Modders
+- Removed `OnlineState.RegisterState`
+
 ## Chat
 
 ### Modders